requirements.txt
google-generativeai
langchain
langchain-community
chromadb
pandas
streamlit
numpy